backend/pyproject.toml
[project]
name = "obsidian-reader"
version = "0.1.0"
description = "A web-based Obsidian vault reader"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-frontmatter>=1.1.0",
    "markdown>=3.7",
    "pygments>=2.18.0",
    "python-multipart>=0.0.17",
    "pyjwt>=2.10.0",
    "passlib[bcrypt]>=1.7.4",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[project.scripts]
obsidian-reader = "obsidian_reader.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_reader"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

